requirements.txt
langchain-openai
fastapi
uvicorn
aiofiles
uvicorn
pdf2image
python-dotenv
python-multipart